# documentation/api/openapi.yaml
openapi: 3.0.3
info:
  title: BudgetApp API
  description: >
    Contratos iniciales para el MVP de BudgetApp (onboarding, presupuestos,
    transacciones, alertas e integraciones bancarias). Esta especificación sirve
    como base para generar SDKs/mocks y evolucionará conforme avancemos en la
    fase de Elaboración.
  version: 0.1.0
  contact:
    name: BudgetApp Team
servers:
  - url: https://api.budgetapp.local
    description: Sandbox local
  - url: https://api.staging.budgetapp.com
    description: Staging
...
paths:
  /api/onboarding:
    post:
      tags: [Onboarding]
      summary: Captura datos iniciales y genera diagnóstico de salud financiera.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OnboardingRequest'
      responses:
        '201':
          description: Diagnóstico generado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OnboardingResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/budgets:
    get:
      tags: [Budgets]
      summary: Obtiene presupuesto por periodo y household.
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: period
          schema:
            type: string
            pattern: '^\d{4}-\d{2}$'
          required: true
          description: Periodo en formato YYYY-MM.
        - in: query
          name: country
          schema:
            type: string
            minLength: 2
            maxLength: 3
          required: false
          description: Código ISO del país cuyo presupuesto se desea consultar.
        - in: query
          name: householdId
          schema:
            type: string
          required: false
          description: Identificador del hogar cuando se trata de un presupuesto familiar.
          required: false
      responses:
        '200':
          description: Presupuesto encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Budget'
    post:
      tags: [Budgets]
      summary: Crea/actualiza presupuesto.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BudgetInput'
      responses:
        '201':
          description: Presupuesto creado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Budget'
        '409':
          description: Ya existe un presupuesto para ese periodo.
  /api/budgets/history:
    get:
      tags: [Budgets]
      summary: Lista todos los presupuestos del usuario según el ámbito seleccionado.
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: scope
          schema:
            type: string
            enum: [personal, household]
          required: false
          description: Determina si se listan presupuestos personales o familiares.
        - in: query
          name: householdId
          schema:
            type: string
          required: false
          description: Filtra por un hogar específico cuando el usuario pertenece a varios.
        - in: query
          name: userId
          schema:
            type: string
          required: false
          description: Solo para agentes/admins, permite consultar el histórico de un usuario concreto.
      responses:
        '200':
          description: Listado de presupuestos.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Budget'
...
  /api/alerts:
    post:
      tags: [Alerts]
      summary: Crea regla de alerta.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlertInput'
      responses:
        '201':
          description: Alerta configurada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Alert'
  /api/integrations/banks/link-token:
    post:
      tags: [Bank Integrations]
      summary: Genera token efímero para iniciar flujo con el proveedor bancario.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                institutionId:
                  type: string
                country:
                  type: string
                  example: CO
              required: [institutionId, country]
      responses:
        '200':
          description: Token generado.
          content:
            application/json:
              schema:
                type: object
                properties:
                  linkToken:
                    type: string
                  expiresIn:
                    type: integer
                    description: Segundos de validez.
  /api/households:
    get:
      tags: [Households]
      summary: Lista los paquetes familiares accesibles para el usuario autenticado.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Catálogo de hogares.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Household'
    post:
      tags: [Households]
      summary: Crea un paquete familiar (roles admin o cs_agent).
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Household'
      responses:
        '201':
          description: Hogar creado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Household'
  /api/households/{id}:
    put:
      tags: [Households]
      summary: Actualiza un paquete familiar existente.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Household'
      responses:
        '200':
          description: Hogar actualizado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Household'
    delete:
      tags: [Households]
      summary: Elimina un paquete familiar.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
      responses:
        '204':
          description: Eliminado.
  /api/households/{id}/members:
    post:
      tags: [Households]
      summary: Agrega o actualiza un miembro del hogar.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HouseholdMember'
      responses:
        '200':
          description: Hogar actualizado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Household'
    delete:
      tags: [Households]
      summary: Elimina un miembro del hogar.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
              required: [userId]
      responses:
        '200':
          description: Hogar actualizado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Household'
...
  /api/reports/kpis:
    get:
      tags: [Reports]
      summary: Devuelve KPIs globales (solo admin).
      security:
        - bearerAuth: []
      responses:
        '200':
          description: KPIs calculados.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KpiSummary'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Error de validación.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            type: string
    OnboardingRequest:
      type: object
      properties:
        country:
          type: string
          minLength: 2
          maxLength: 3
        ingresosMensuales:
          type: number
        gastosFijos:
          type: number
        deudas:
          type: array
          items:
            $ref: '#/components/schemas/Debt'
        metas:
          type: array
          items:
            $ref: '#/components/schemas/Goal'
      required: [country, ingresosMensuales, gastosFijos]
    OnboardingResponse:
      type: object
      properties:
        scoreSalud:
          type: integer
          minimum: 0
          maximum: 100
        recomendaciones:
          type: array
          items:
            type: string
    BudgetInput:
      type: object
      properties:
        householdId:
          type: string
          nullable: true
          description: Asociar a un hogar para presupuestos compartidos.
        scope:
          type: string
          enum: [personal, household]
          description: Determina si el presupuesto afecta al ámbito personal o familiar.
        period:
          type: string
        currency:
          type: string
        country:
          type: string
          minLength: 2
          maxLength: 3
        frequency:
          type: string
          enum: [monthly, annual]
          default: monthly
          description: Periodicidad del presupuesto; por default es mensual.
        userId:
          type: string
          description: Sólo agentes/admins pueden usarlo para guardar presupuestos de otro usuario.
        userEmail:
          type: string
          format: email
          description: Alternativa para que CS Agents busquen al usuario objetivo por correo.
        startBalance:
          type: number
          description: Saldo inicial (Start Balance) definido en Start Here.
        startWithMoney:
          type: boolean
          description: Indica si el periodo arranca con dinero ya asignado.
        categories:
          type: array
          items:
            $ref: '#/components/schemas/BudgetCategory'
      required: [period, currency, country, categories]
    Budget:
      allOf:
        - $ref: '#/components/schemas/BudgetInput'
        - type: object
          properties:
            id:
              type: string
            householdId:
              type: string
            totals:
              type: object
              properties:
                asignado:
                  type: number
                gastado:
                  type: number
    BudgetCategory:
      type: object
      properties:
        categoryId:
          type: string
        asignado:
          type: number
        gastado:
          type: number
        subcategories:
          type: array
          items:
            $ref: '#/components/schemas/BudgetSubcategory'
    BudgetSubcategory:
      type: object
      properties:
        id:
          type: string
        label:
          type: string
        asignado:
          type: number
        startWithMoney:
          type: boolean
          readOnly: true
      required: [categoryId, asignado]
...
    AlertInput:
      type: object
      properties:
        type:
          type: string
          enum: [budget_over, transaction_large, goal_at_risk]
        threshold:
          type: number
        channels:
          type: array
          items:
            type: string
            enum: [push, email, whatsapp]
      required: [type, channels]
    Alert:
      allOf:
        - $ref: '#/components/schemas/AlertInput'
        - type: object
          properties:
            id:
              type: string
            createdAt:
              type: string
              format: date-time
    Debt:
      type: object
      properties:
        name:
          type: string
        balance:
          type: number
        interestRate:
          type: number
    Goal:
      type: object
      properties:
        name:
          type: string
        targetAmount:
          type: number
        targetDate:
          type: string
          format: date
...
    HouseholdMember:
      type: object
      properties:
        userId:
          type: string
        role:
          type: string
          enum: [owner, member]
        sharePercent:
          type: number
      required: [userId]
    Household:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        planType:
          type: string
        billingUserId:
          type: string
        members:
          type: array
          items:
            $ref: '#/components/schemas/HouseholdMember'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, planType, billingUserId]
...
    KpiSummary:
      type: object
      properties:
        totalUsers:
          type: integer
        totalClients:
          type: integer
        totalAgents:
          type: integer
        activeHouseholds:
          type: integer
        openTickets:
          type: integer
        averageHouseholdBudget:
          type: number
...